First pass implementing Agency page
It's not exactly pixel perfect but attempt to recreate the Agency page
using only Flow layout CSS.

I avoid margins in favour of padding and so have to reset some of the
browser defaults though I leave the default <p> tag margins alone.

The grey background of the <main> element won't stretch to fill the
viewport on very tall screens but this was ambiguous in the spec so I
left it for now.